Resolves FE-7729
The issue is that Qt would lose the 'active Window' pointer, which in normal operation indicates that the application no longer has focus.  However, in 3ds Max because the Qt window is hosted within a Win32 window, in a C# application, the normal operation does not appear to work.  If there is no active window, then the Qt Shortcut system will simply ignore keystrokes.

however the fix was a lot more involved than expected.

Firstly, during every keystroke, Qt would recieve an ACTIVATEDAPP message, telling Qt that it no longer has focus.  Where this message comes from is still unknown, I suspect its posted by the parent app in response to Qt taking focus.  To fix this, I installed a filter on the Qt message pump, and discard all of these messages when a Qt dialog is focused.  The side-effects of this fix are unknown, however it will not result in incorrect keystrokes being piped to Qt, as the application will still only receive keystrokes while active.

Also, frequently during operation, Qt would reset the active Window to null.  For example, when a modal dialog exits (eg, after "Add Graph" dlg), the dialogs cleanup code would recognize the modal window as the app window and set the activeWindow to null.  It is unknown how this is expected to work in a 'normal' Qt environment, I suspect that the dialog would not become the active window, or somehow the active window reverts to the application when it becomes null.  To fix this issue, I added a OnKeyDown filter to the message pump, and for every keydown detected it resets the active window to the window of the focused dialog item.

Also noted: Undo/Redo filter was not consistently trapping keystrokes.  When in a child window, the undo/redo keystrokes could be directed to them and not be picked up by the keyfilter I was using to trap the key sequence.  To fix this, I took the undo/redo QActions, and added them to the window QWidget (so the QActions would get triggered whenever this window is active).  I then removed the explicit key trappings.  This method is more consistent with fabric's implementation of QActions anyway.

Behaviour appears to be working as intended.

Tested:
Ctrl-A shortcut in GraphView
Ctrl-A shortcut in Value Editor (ensure that shortcuts go to correct widgets)

Multiple DFG Editors open simultaneously.
Closing/Reopening DFG Editor
Undo/Redo in Text Editor, main window, and function editor window, different simultaneous DFG editors.